Create README.md
##Abstract
The purpose of this project is the electrical and mechanical design of a transmission line. First, a suitable conductor must be selected based on characteristics such as passing current, loss voltage, short circuit current, corona phenomenon.
The next step is choosing the right wire guard for the transmission line. Then the insulation calculations should be obtained, including overvoltages caused by switching, lightning, power frequency and pollution, and then the number of insulators and air gap should be determined. Then the mechanical calculations of the transmission line are done and after determining the loading table, the calculations related to wire tension and span or the distance between two towers are determined.